Pulling the trigger on ArcStr
This commit fully converts `gramatika` and `gramatika-macro` from the
lifetime-heavy String / &str implementation to ArcStr / Substr. A few
major consequences of this change:

* The need for lifetime annotations is completely eliminated (!)
* The need for `unsafe` fuckery internally is completely eliminated (!!)

This will require a fairly substantial migration effort for nontrivial
consuming codebases, which will mostly involve tediously removing
lifetime annotations. But on the bright side, no more lifetimes!

The perf story is a bit mixed. The benchmarks are indicating negligible
overhead on the Lexer, with test cases only showing an uptick of a few
percentage points. The impact on parsing has been much more substantial.
This requires further investigation, but I'm optimistic that there's
room for optimization here without losing much (if anything) in the way
of ergonomics.