Fix lost tokens on conversion failure (#8)
Previously, the `consume_as` and `upgrade_last` implementations in
`ParseStream` worked as follows:

* Pop the last token from the buffer
* Check the condition indicated by the function argument
  - If it passes:
    - Convert the token with the provided function
    - Push the converted token back onto the buffer
    - return an `Ok` with the converted token
  - Else:
    - return an `Err` indicating the condition failure

The fact that the `else` branch did _not_ push the original token back
onto the buffer before returning would cause that token to be lost
completely, which could cause unexpected issues if the user intended to
handle the returned `Err`. This PR corrects that oversight by restoring
the original token to the buffer before returning the `Err`.