`substr-source` feature and concrete `Lexer` implementation (#9)
## New Cargo feature: `substr-source`

When enabled, this replaces `arcstr::ArcStr` with `arcstr::Substr` for
all fields and methods that represent the original source text of a
document being scanned or parsed. A root-level `SourceStr` type alias
has also been added to make it easier to refer to this type regardless
of whether the feature is enabled or not.

This feature enables optimizations for some niche advanced use cases
that are currently being explored in [`wgsl_parser`]. Namely, this
allows a completely different lexer/parser combinations to operate on
subsets of a document, without needing to allocate a new `ArcStr` for
each subset.

In the future, this could be leveraged to extend Gramatika's utility to
languages like XML that require a stateful, context-sensitive lexer that
exhibits different scanning behavior depending on its current state, but
that would require some additional effort and API design which is out of
scope for this particular update.

[`wgsl_parser`]: https://crates.io/crates/wgsl_parser

## Concrete `Lexer` implementation and deprecation of `#[derive(Lexer)]`

Additionally, this update deprecates the `Lexer` derive macro, replacing
it with a new concrete `TokenStream` type provided by the crate. This
follows the same convention used by the `parse` module, where the
`ParseStream` type is provided as an optional, concrete implementation
of the `ParseStreamer` trait.

The `Lexer` derive was often a bit awkward to use (and extremely
unhygienic), because it emitted a phantom `struct Lexer` type which
couldn't be directly inspected, renamed, or otherwise modified by the
user.

To migrate, simply remove the `#[derive(Lexer)]` attribute from your
token type, and replace all references to the previously generated
`Lexer` struct with `gramatika::TokenStream<T>`, where `T` is your token
type.